Fix build
So it turns out that the gloo-envoy-wrapper container we have uses a
truly magical construction: A glibc binary in a musl-based docker image.
They made that work by shipping a glibc compat layer and adjusting the
start.
Of course, since this thing will be dlopen'd into that process, we need
to reproduce this. That means building on the same ubuntu version (so
that the glibc we use is the same) _but_ statically linking libc++ of a
matching version.